Moved Instabug implemenation to SimplifiedServices
