+++ simplified-app-shared/src/main/java/org/nypl/simplified/app/services/SimplifiedServices.kt
@@ -9,6 +9,8 @@ import androidx.core.content.ContextCompat
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.google.common.base.Preconditions
 import com.google.common.util.concurrent.ListeningScheduledExecutorService
+import com.instabug.library.Instabug
+import com.instabug.library.invocation.InstabugInvocationEvent
 import com.io7m.jfunctional.Option
 import com.io7m.jfunctional.OptionType
 import com.io7m.jfunctional.Some
@@ -32,12 +34,7 @@ import org.nypl.simplified.analytics.api.Analytics
 import org.nypl.simplified.analytics.api.AnalyticsConfiguration
 import org.nypl.simplified.analytics.api.AnalyticsEvent
 import org.nypl.simplified.analytics.api.AnalyticsType
-import org.nypl.simplified.app.AdobeDRMServices
-import org.nypl.simplified.app.Bugsnag
-import org.nypl.simplified.app.BundledContentResolver
-import org.nypl.simplified.app.NetworkConnectivityType
-import org.nypl.simplified.app.R
-import org.nypl.simplified.app.ScreenSizeInformationType
+import org.nypl.simplified.app.*
 import org.nypl.simplified.app.SimplifiedNetworkConnectivity
 import org.nypl.simplified.app.catalog.CatalogBookBorrowStrings
 import org.nypl.simplified.app.catalog.CatalogBookRevokeStrings
@@ -109,6 +106,7 @@ import java.io.File
 import java.io.FileNotFoundException
 import java.io.IOException
 import java.net.ServerSocket
+import java.util.Properties
 import java.util.ServiceLoader
 import java.util.concurrent.ExecutorService
 
@@ -173,6 +171,7 @@ class SimplifiedServices private constructor(
      */
 
     const val CURRENT_DATA_VERSION = "v4.0"
+    const val SIMPLYE = "org.nypl.simplified.simplye"
 
     private data class Directories(
       val directoryPrivateBaseVersioned: File,
@@ -201,6 +200,9 @@ class SimplifiedServices private constructor(
 
       publishEvent(strings.bootingBugsnag)
       this.initBugsnag(context, Bugsnag.getApiToken(assets))
+
+      publishEvent(strings.initializingInstabug)
+      this.configureInstabug(context)
 
       publishEvent(strings.bootingAdobeDRM)
       val adobeDRM =
@@ -779,6 +781,33 @@ class SimplifiedServices private constructor(
       }
 
       return documentsBuilder.build()
+    }
+
+    /**
+     * Currently Instabug is available for SimplyE debug builds only
+     */
+    private fun configureInstabug(context: Context) {
+      if (BuildConfig.DEBUG) {
+        if (Simplified.application.packageName == SIMPLYE) {
+
+          try {
+            val inputStream = context.assets.open("instabug.conf")
+            val p = Properties()
+            p.load(inputStream)
+            val instabugToken = p.getProperty("instabug.token")
+
+            if (instabugToken.isNullOrBlank()) {
+              throw IOException("instabug token not found!")
+            }
+            Instabug.Builder(Simplified.application, instabugToken)
+                    .setInvocationEvents(InstabugInvocationEvent.SHAKE, InstabugInvocationEvent.SCREENSHOT)
+                    .build()
+          } catch (e: java.lang.Exception) {
+            this.logger.debug("Error intializing Instabug", android.os.Process.myPid())
+            e.printStackTrace()
+          }
+        }
+      }
     }
   }
